Time: 47 ms (78.29%), Space: 27.6 MB (63.73%) - LeetHub

+++ 704-binary-search/704-binary-search.cpp
@@ -1,12 +1,17 @@
 class Solution {
 public:
     int search(vector<int>& nums, int target) {
+       
+        
+        
+        
+        
         int left = 0;
         int right = nums.size() - 1;
         
        while(left <= right){
             
-            int m = left + (right - left) / 2;
+            int16_t m = left + (right - left) / 2;
             
             if(nums[m] == target){
                 return m;
@@ -19,6 +24,8 @@ public:
             }
         }
         return -1;
+        
+        
         
     }
 };
